+++ apps/native_poc/src/native_video_client_tcp_control.hpp
@@ -1,9 +1,8 @@
 #pragma once
 
-#include <winsock2.h>
-
 #include <cstdint>
 
+#include "native_socket.hpp"
 #include "native_video_client_shared_core.hpp"
 
 namespace remote60::native_poc {
@@ -24,8 +23,8 @@ struct TcpControlResponse {
   ControlInputAckMessage inputAck{};
 };
 
-bool send_tcp_control_action(SOCKET controlSock, const ControlOutboundAction& action);
-bool recv_tcp_control_response(SOCKET controlSock, const ControlOutboundAction& action, TcpControlResponse* out);
-bool execute_tcp_control_action(SOCKET controlSock, const ControlOutboundAction& action, TcpControlResponse* out);
+bool send_tcp_control_action(SocketHandle controlSock, const ControlOutboundAction& action);
+bool recv_tcp_control_response(SocketHandle controlSock, const ControlOutboundAction& action, TcpControlResponse* out);
+bool execute_tcp_control_action(SocketHandle controlSock, const ControlOutboundAction& action, TcpControlResponse* out);
 
 }  // namespace remote60::native_poc